Breakout grids to module
